Command Prefixes and Wild Cards!
I've added the ability to set command prefixes. Also, I'm allowing
wild card characters to be used in either the command prefix or
the arguments settings.

This replaces the need for the SurpressDirectoryArgument setting,
as this is more flexible

I've also replaced MacWinRunDirectly with runDirectly, which now also
applies to Linux.